Title: 20대의 평균나이, Time: -1ms, Memory: -1MB, Status: Runtime Error - CodeTree

+++ 240902/20대의 평균나이/average-age-of-20.py
@@ -7,7 +7,7 @@ def main():
         if age >= 30:
             break
         ages.append(age)
-    print('{:.2f}'.format(sum(ages)/len(ages)))
+    print('{:.2f}'.format(float(sum(ages)/len(ages))))
